Added cut down pure pursuit
It is derivied from Jiaming's version but with all of the fat cut out. The issue is that relative position and orientation were never computer. I will prob reimplment this with my c++ control code as a base so I have simpler time plotting and logging error. Goog try though Jiaming. Comments were on points.